+++ EssentialFeedTests/FeedCache/CacheFeedImageDataUseCaseTests.swift
@@ -1,0 +1,39 @@
+//
+//  CacheFeedImageDataUseCaseTests.swift
+//  EssentialFeedTests
+//
+//  Created by Firdavs Bagirov on 03/10/22.
+//
+
+import XCTest
+import EssentialFeed
+
+class CacheFeedImageDataUseCaseTests: XCTestCase {
+    
+    func test_init_doesNotMessageStoreUponCreation() {
+        let (_, store) = makeSUT()
+        
+        XCTAssertTrue(store.receivedMessages.isEmpty)
+    }
+    
+    func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
+        let (sut, store) = makeSUT()
+        let url = anyURL()
+        let data = anyData()
+        
+        sut.save(data, for: url) { _ in }
+        
+        XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
+    }
+    
+    // MARK: - Helpers
+    
+    private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: LocalFeedImageDataLoader, store: FeedImageDataStoreSpy) {
+        let store = FeedImageDataStoreSpy()
+        let sut = LocalFeedImageDataLoader(store: store)
+        trackForMemoryLeak(store, file: file, line: line)
+        trackForMemoryLeak(sut, file: file, line: line)
+        return (sut, store)
+    }
+    
+}
